requirements.txt
fastapi
uvicorn
langchain
langchain-aws
langgraph
boto3
python-dotenv
pymupdf
chromadb
langchain-text-splitters
langchain-community
langchain-chroma